chess-outcome-predictor/requirements.txt
pandas
numpy
scikit-learn
python-chess
matplotlib